Refactor container context and remove HandleServerMessageGuard
