+++ src/App.js
@@ -3,7 +3,6 @@ import { Routes, Route, Outlet, useLocation } from 'react-router-dom';
 // Import i18next and set up translations
 import { I18nextProvider } from 'react-i18next';
 
-import handleServerMessagesGuard from './hoc/handleServerMessagesGuard';
 import i18n from './i18n';
 import useCheckURL from './hooks/useCheckURL';
 import { CredentialsProvider } from './context/CredentialsContext';
@@ -15,7 +14,7 @@ import HandlerNotification from './components/Notifications/HandlerNotification'
 import Snowfalling from './components/ChristmasAnimation/Snowfalling';
 import Spinner from './components/Shared/Spinner';
 
-import { ContainerContextProvider } from './context/ContainerContext';
+import { withContainerContext } from './context/ContainerContext';
 
 import UpdateNotification from './components/Notifications/UpdateNotification';
 import CredentialDetails from './pages/Home/CredentialDetails';
@@ -80,7 +79,6 @@ const SendCredentials = React.lazy(() => import('./pages/SendCredentials/SendCre
 const Settings = React.lazy(() => import('./pages/Settings/Settings'));
 const VerificationResult = React.lazy(() => import('./pages/VerificationResult/VerificationResult'));
 
-
 function App() {
 	const location = useLocation();
 	const [url, setUrl] = useState(window.location.href);
@@ -123,63 +121,62 @@ function App() {
 			window.location.href = homeURL;
 		}
 	};
+
 	return (
 		<I18nextProvider i18n={i18n}>
 			<CredentialsProvider>
-				<ContainerContextProvider>
-					<Snowfalling />
-					<Suspense fallback={<Spinner />}>
-						<HandlerNotification />
-						<UpdateNotification />
-						<Routes>
-							<Route element={
-								<PrivateRoute>
-									<Layout>
-										<Suspense fallback={<Spinner />}>
-											<PrivateRoute.NotificationPermissionWarning />
-											<FadeInContentTransition appear reanimateKey={location.pathname}>
-												<Outlet />
-											</FadeInContentTransition>
-										</Suspense>
-									</Layout>
-								</PrivateRoute>
-							}>
-								<Route path="/settings" element={<Settings />} />
-								<Route path="/" element={<Home />} />
-								<Route path="/credential/:credentialId" element={<Credential />} />
-								<Route path="/credential/:credentialId/history" element={<CredentialHistory />} />
-								<Route path="/credential/:credentialId/details" element={<CredentialDetails />} />
-								<Route path="/history" element={<History />} />
-								<Route path="/history/:historyId" element={<HistoryDetail />} />
-								<Route path="/add" element={<AddCredentials />} />
-								<Route path="/send" element={<SendCredentials />} />
-								<Route path="/verification/result" element={<VerificationResult />} />
-								<Route path="/cb/*" element={<Home />} />
-							</Route>
-							<Route element={
-								<FadeInContentTransition reanimateKey={location.pathname}>
-									<Outlet />
-								</FadeInContentTransition>
-							}>
-								<Route path="/login" element={<Login />} />
-								<Route path="/login-state" element={<LoginState />} />
-								<Route path="*" element={<NotFound />} />
-							</Route>
-						</Routes>
-						{showSelectCredentialsPopup &&
-							<SelectCredentialsPopup isOpen={showSelectCredentialsPopup} setIsOpen={setShowSelectCredentialsPopup} setSelectionMap={setSelectionMap} conformantCredentialsMap={conformantCredentialsMap} verifierDomainName={verifierDomainName} />
-						}
-						{showPinInputPopup &&
-							<PinInputPopup isOpen={showPinInputPopup} setIsOpen={setShowPinInputPopup} />
-						}
-						{showMessagePopup &&
-							<MessagePopup type={typeMessagePopup} message={textMessagePopup} onClose={() => setMessagePopup(false)} />
-						}
-					</Suspense>
-				</ContainerContextProvider>
+				<Snowfalling />
+				<Suspense fallback={<Spinner />}>
+					<HandlerNotification />
+					<UpdateNotification />
+					<Routes>
+						<Route element={
+							<PrivateRoute>
+								<Layout>
+									<Suspense fallback={<Spinner />}>
+										<PrivateRoute.NotificationPermissionWarning />
+										<FadeInContentTransition appear reanimateKey={location.pathname}>
+											<Outlet />
+										</FadeInContentTransition>
+									</Suspense>
+								</Layout>
+							</PrivateRoute>
+						}>
+							<Route path="/settings" element={<Settings />} />
+							<Route path="/" element={<Home />} />
+							<Route path="/credential/:credentialId" element={<Credential />} />
+							<Route path="/credential/:credentialId/history" element={<CredentialHistory />} />
+							<Route path="/credential/:credentialId/details" element={<CredentialDetails />} />
+							<Route path="/history" element={<History />} />
+							<Route path="/history/:historyId" element={<HistoryDetail />} />
+							<Route path="/add" element={<AddCredentials />} />
+							<Route path="/send" element={<SendCredentials />} />
+							<Route path="/verification/result" element={<VerificationResult />} />
+							<Route path="/cb/*" element={<Home />} />
+						</Route>
+						<Route element={
+							<FadeInContentTransition reanimateKey={location.pathname}>
+								<Outlet />
+							</FadeInContentTransition>
+						}>
+							<Route path="/login" element={<Login />} />
+							<Route path="/login-state" element={<LoginState />} />
+							<Route path="*" element={<NotFound />} />
+						</Route>
+					</Routes>
+					{showSelectCredentialsPopup &&
+						<SelectCredentialsPopup isOpen={showSelectCredentialsPopup} setIsOpen={setShowSelectCredentialsPopup} setSelectionMap={setSelectionMap} conformantCredentialsMap={conformantCredentialsMap} verifierDomainName={verifierDomainName} />
+					}
+					{showPinInputPopup &&
+						<PinInputPopup isOpen={showPinInputPopup} setIsOpen={setShowPinInputPopup} />
+					}
+					{showMessagePopup &&
+						<MessagePopup type={typeMessagePopup} message={textMessagePopup} onClose={() => setMessagePopup(false)} />
+					}
+				</Suspense>
 			</CredentialsProvider>
 		</I18nextProvider>
 	);
 }
 
-export default withSessionContext(handleServerMessagesGuard(App));
+export default withSessionContext(withContainerContext(App));
